pyproject.toml
[project]
name = "mcp-katas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.122.0",
    "mcp>=1.22.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
